:bug: 1. 隐藏交易大赛open order 实时价格的人民币价格 2. 交易大赛banner在Activity recreate后不显示修复 3. 交易界面下拉框文字分行

+++ app/src/main/java/com/cybexmobile/graphene/chain/TradeHistory.java
@@ -13,7 +13,7 @@ public class TradeHistory {
     public Receive receives;
 
     public class Fee {
-        public int amount;
+        public long amount;
         public String asset_id;
     }
 
@@ -24,12 +24,12 @@ public class TradeHistory {
 
 
     public class Pay {
-        public int amount;
+        public long amount;
         public String asset_id;
     }
 
     public class Receive {
-        public int amount;
+        public long amount;
         public String asset_id;
     }
